Add 2025-07-13 FE요구사항 추가

+++ src/follows/follows.service.ts
@@ -1,9 +1,12 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
-import { Follow } from './entities/follow.entity';
-import { Follow as FollowSchema, FollowDocument } from '../schemas/follow.schema';
+import {
+  FollowDocument,
+  Follow as FollowSchema,
+} from '../schemas/follow.schema';
 import { FollowInput } from './dto/follow.input';
+import { Follow } from './entities/follow.entity';
 
 @Injectable()
 export class FollowsService {
@@ -12,10 +15,12 @@ export class FollowsService {
     private readonly followModel: Model<FollowDocument>,
   ) {}
 
-  async create(input: FollowInput): Promise<Follow> {
+  async create(input: FollowInput, userId: string): Promise<Follow> {
     const follow = new this.followModel({
       followerId: input.followerId,
       followingId: input.followingId,
+      createdBy: userId,
+      updatedBy: userId,
       // status 등 필요한 필드 추가
     });
     const saved = await follow.save();
@@ -28,7 +33,7 @@ export class FollowsService {
 
   async findAll(): Promise<Follow[]> {
     const follows = await this.followModel.find();
-    return follows.map(f => ({
+    return follows.map((f) => ({
       id: f._id ? String(f._id) : '',
       followerId: f.followerId,
       followingId: f.followingId,
@@ -45,12 +50,17 @@ export class FollowsService {
     };
   }
 
-  async update(id: string, input: FollowInput): Promise<Follow | undefined> {
+  async update(
+    id: string,
+    input: FollowInput,
+    userId: string,
+  ): Promise<Follow | undefined> {
     const f = await this.followModel.findByIdAndUpdate(
       id,
       {
         followerId: input.followerId,
         followingId: input.followingId,
+        updatedBy: userId,
       },
       { new: true },
     );
@@ -62,8 +72,22 @@ export class FollowsService {
     };
   }
 
-  async remove(id: string): Promise<boolean> {
+  async remove(id: string, userId: string): Promise<boolean> {
+    const f = await this.followModel.findByIdAndUpdate(
+      id,
+      { updatedBy: userId },
+      { new: true },
+    );
+    if (!f) return false;
     const res = await this.followModel.deleteOne({ _id: id });
     return res.deletedCount > 0;
   }
-} 
+
+  async isFollowing(followerId: string, followingId: string): Promise<boolean> {
+    const follow = await this.followModel.findOne({
+      followerId,
+      followingId,
+    });
+    return !!follow;
+  }
+}
